Refactor extract PasswordEdit to 'gui' namespace
