+++ securedrop_client/gui/base/__init__.py
@@ -18,6 +18,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 # Import classes here to make possible to import them from securedrop_client.gui.base
 from securedrop_client.gui.base.buttons import SDPushButton  # noqa: F401
+from securedrop_client.gui.base.inputs import PasswordEdit  # noqa: F401
 from securedrop_client.gui.base.misc import (  # noqa: F401
     SecureQLabel,
     SvgLabel,
