+++ securedrop_client/gui/widgets.py
@@ -77,6 +77,7 @@ from securedrop_client.db import (
 )
 from securedrop_client.export import ExportError, ExportStatus
 from securedrop_client.gui.base import (
+    PasswordEdit,
     SDPushButton,
     SecureQLabel,
     SvgLabel,
@@ -1713,34 +1714,6 @@ class LoginErrorBar(QWidget):
     def clear_message(self) -> None:
         self.error_status_bar.setText("")
         self.hide()
-
-
-class PasswordEdit(QLineEdit):
-    """
-    A LineEdit with icons to show/hide password entries
-    """
-
-    def __init__(self, parent: QDialog) -> None:
-        self.parent = parent
-        super().__init__(self.parent)
-
-        self.visibleIcon = load_icon("eye_visible.svg")
-        self.hiddenIcon = load_icon("eye_hidden.svg")
-
-        self.setEchoMode(QLineEdit.Password)
-        self.togglepasswordAction = self.addAction(self.hiddenIcon, QLineEdit.TrailingPosition)
-        self.togglepasswordAction.triggered.connect(self.on_toggle_password_Action)
-        self.password_shown = False
-
-    def on_toggle_password_Action(self) -> None:
-        if not self.password_shown:
-            self.setEchoMode(QLineEdit.Normal)
-            self.password_shown = True
-            self.togglepasswordAction.setIcon(self.visibleIcon)
-        else:
-            self.setEchoMode(QLineEdit.Password)
-            self.password_shown = False
-            self.togglepasswordAction.setIcon(self.hiddenIcon)
 
 
 class SenderIcon(QWidget):
